+++ src/main/java/com/inventory/Inventory/repository/ColorRepository.java
@@ -3,5 +3,5 @@ package com.inventory.Inventory.repository;
 import com.inventory.Inventory.model.Colors;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColorRepository extends JpaRepository<Colors, Long> {
+public interface ColorRepository extends JpaRepository<Colors, Integer> {
 }
